Add GANUtilities and move CreateNoiseSamples method
