+++ src/CntkCatalyst.Examples/CntkCatalyst.Examples.csproj
@@ -87,6 +87,7 @@
     <Compile Include="DeepLearningFrancoisChollet\Ch_51_Introduction_To_Convnets.cs" />
     <Compile Include="DeepLearningFrancoisChollet\Ch_35_Classifying_Movie_Reviews.cs" />
     <Compile Include="DeepLearningFrancoisChollet\Ch_52_Using_Convnets_With_Small_Datasets.cs" />
+    <Compile Include="GenerativeModels\GANUtilities.cs" />
     <Compile Include="GenerativeModels\GAN_Basic.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
